Verify crtc against previous config.
When selecting a crtc for an output, it doesn't check for already attached
outputs in check_crtc_for_output(), and so may select an crtc that is already
in use.
E.g. when changing from a cloned mode the displays won't be split up to
different crtcs, but the same crtc will be used, changing the resolution for
the unspecified display as well.