xrandr: make --query really the default option
The manpage states that "--query" is the default action when nothing
specific is requested. However, some options such as "-display" or
"--screen" lead to do nothing by default. This makes sure that unless
something specific is requested, query is done.

This also restores the behaviour of --q1 to the same one as the old
xrandr: also display the info on rotation and reflection by default.

Signed-off-by: Éric Piel <eric.piel@tremplin-utc.net>
Signed-off-by: Matthias Hopf <mhopf@suse.de>